refactor MigrationFile, DatabaseDriver, Stubs
